get host.zig#roc_fx_fileReadBytes() working
thanks and credit to https://github.com/bhansconnect for most of this

for now InternalFile.ReadErr#Unrecognized is commented out so that the
structure sizes match. this was likely the main reason my attempts
didn't work. this is a temparary workaround until
host.zig/RocResult#payload becomes a union T,E

Signed-off-by: Travis Staloch <twostepted@gmail.com>